// src/app/dashboard/campaigns/new-prebuild-template/components/types.ts
export interface TemplateFormState {
  templateName: string;
  advertiser: string;
  includeLocation: boolean;
  location: string;
  library: string;
  date: string;
  makeFilter: string[];
  yearStart: string;
  yearEnd: string;
  authorize: boolean;
  campaignName: string;
  campaignStatus: string;
  budget: string;
  networks: string;
  enhancedCpc: string;
  mobileBidModifier: string;
  adRotation: string;
  negativeKeywords: string[];
  adGroupName: string;
  adGroupStatus: string;
  finalUrl: string;
  maxCpcBid: string;
  setMaxCpcOnCreate: boolean;
} 